+++ django_veo_admin_api/changelist.py
@@ -15,7 +15,7 @@ from django.utils import timezone
 from django.utils.translation import gettext as _
 
 from django_veo_admin_api.constants import ShowFacets
-from django_veo_admin_api.exceptions import AdminValidationError, DisallowedModelAdminLookup
+from django_veo_admin_api.core.exceptions import AdminValidationError, DisallowedModelAdminLookup
 from django_veo_admin_api.filters import build_filter_spec
 from django_veo_admin_api.utils.lookup import (
     field_name_for_display,